fleur: sysbta: Implement all existing audio HAL versions
This is needed because libaudiohal.so can only interact with one HAL
version at a time (loading only one of the libaudiohal@version.so
implementations), so only the interface that matches the stock HAL
version will be used. libaudiohal decides which version to use by
looking through the interfaces whose variant is `default` and picking
the highest versioned one, so our hal will not interfere in the process
(because we use `sysbta` as our variant).


Former-commit-id: bedf522dc684d2b795dad668c1f53db5b4cd7d15
Former-commit-id: 6c44220ee0b0872c4813f52cb85e11bdd4086bf3